refactor(settings): move timezone and other choice fields into constants
This commit moves the generation of a list of timezones and other
choice fields, such as user roles and reporting source types, from
settings into constants.
